refact: letters -> letter 수정
