+++ src/main/java/tmfm/tomefromme/global/JasyptConfig.java
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class JasyptConfig {
 
-    private final String key = "MY_JASYPT_KEY";
+    private final String key = "TMFM_JASYPT_KEY";
 
     @Bean
     public StringEncryptor jasyptStringEncryptor() {
