+++ src/main/java/tmfm/tomefromme/web/letter/controller/LettersController.java
@@ -41,8 +41,4 @@ public class LettersController {
         }
     }
 
-    @PostMapping("/letter")
-    public void createLetter(@RequestParam Letter letter, @RequestParam List<MultipartFile> multipartFileList) {
-        lettersService.createLetter(letter, multipartFileList);
-    }
 }
