+++ src/main/java/tmfm/tomefromme/web/letter/repository/LetterRepository.java
@@ -1,0 +1,9 @@
+package tmfm.tomefromme.web.letter.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import tmfm.tomefromme.domain.letter.entity.Letter;
+
+@Repository
+public interface LetterRepository extends JpaRepository<Letter, Long> {
+}
